WIP: Added necessary Suspense component
