+++ src/styles/AuthPage.module.css
@@ -1,0 +1,104 @@
+.container {
+  width: 100%;
+  height: 100%;
+  position: relative;
+  box-sizing: border-box;
+}
+
+.imageBox {
+  width: 520px;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  right: 0;
+  background-color: #f6f6f6;
+}
+
+.image {
+  object-fit: contain;
+}
+
+.content {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  z-index: 2;
+  background-color: var(--white);
+  box-shadow: 0px 8px 28px 0px rgba(0, 0, 0, 0.1);
+  transition: right 0.3s ease-in-out;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.contentLogin {
+  right: 520px;
+}
+
+.contentRegister {
+  right: 0;
+}
+
+.header {
+  padding: 2.5rem 2.5rem 0;
+  display: flex;
+  flex-direction: column;
+  gap: 0.625rem;
+}
+
+.buttonsBox {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+}
+
+.authButton {
+  font-family: var(--font-poppins);
+  font-size: 1.5rem;
+  font-weight: 700;
+  line-height: 1.625rem;
+  color: #a6a6a6;
+  cursor: pointer;
+  transition: color 0.3s ease-in-out;
+}
+
+.authButton:hover {
+  color: #6e6e6e;
+}
+
+.authButtonActive {
+  color: #0c0c0c !important;
+}
+
+.line {
+  height: 4px;
+  background-color: #000000;
+  position: relative;
+  transition: width 0.3s ease-in-out, left 0.3s ease-in-out;
+}
+
+.lineLogin {
+  width: 65.31px;
+  left: 0;
+}
+
+.lineRegister {
+  width: 96.05px;
+  left: calc(100% - 96.05px);
+}
+
+.wrapper {
+  position: relative;
+  flex-grow: 1;
+  display: flex;
+  transition: left 0.3s ease-in-out;
+}
+
+.wrapperLogin {
+  left: 0;
+}
+
+.wrapperRegister {
+  left: -400px;
+}
